fix folder mentions
